requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.5.4
gunicorn==23.0.0
